some clean up

+++ src/Hydroinfo_MUSIC.cpp
@@ -693,7 +693,7 @@ void Hydroinfo_MUSIC::readHydroData(int whichHydro, int nskip_tau_in) {
         // create the index map
         long long ncells = (itaumax + 1)*ixmax*ixmax*ietamax;
         idx_map_.resize(ncells, 0);
-        for (int i = 0; i < lattice_3D_ideal.size(); i++) {
+        for (int i = 0; i < static_cast<int>(lattice_3D_ideal.size()); i++) {
             const auto cell_i = lattice_3D_ideal[i];
             int cell_idx = (
                 (  (cell_i.itau*ietamax + cell_i.ieta)*ixmax
